backend/sparky/requirements.txt
# Sparky Agent Dependencies
# Streaming agent service with LangChain/LangGraph

# LangChain ecosystem - required for streaming agent
langgraph>=1.1.2
langchain==1.2.12
langchain-aws==1.4.1
langgraph-checkpoint-aws==1.0.6

# MCP tool support - for loading cached MCP tools
langchain-mcp-adapters==0.2.1

# Tavily web search tool
langchain-tavily==0.2.17

# Web framework
uvicorn[standard]==0.41.0
fastapi==0.135.1

# HTTP client for MCP tool calls
httpx==0.28.1

# JWT token validation
PyJWT==2.10.1

# Image processing for attachments
Pillow==11.0.0

# Browser automation via CDP WebSocket
playwright==1.58.0
bedrock-agentcore==1.4.3

# Package installer (used by MCP server management)
uv==0.10.8